be_fastAPI/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
apify_client==1.8.1
apify_shared==1.1.2
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
more-itertools==10.5.0
motor==3.6.0
numpy==2.1.3
pandas==2.2.3
pydantic==2.10.2
pydantic_core==2.27.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
python-dotenv==0.19.1
